import common worker and job stucts from gostockd/common (gostock)
